+++ DAL/Request/Dto/NewUser.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Model.Models
+namespace Models.Request.Dto
 {
     public class NewUser
     {
