+++ Model/Domain/AdminContext.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Model.Models
+namespace DAL.Domain
 {
     public class AdminContext : DbContext
     {
