+++ Model/Domain/BaseEntity.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace Model.Models
+namespace DAL.Domain
 {
     public class BaseEntity
     {
