+++ WebApi.SocialNetWorkAdministration/Controllers/AdminController.cs
@@ -3,8 +3,8 @@ using BL.AdminLogic;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Model.Models;
-using Model.Models.Dto;
+using DAL.Domain;
+using DAL.Domain.Dto;
 using System;
 using System.Collections.Generic;
 using System.Linq;
