+++ WebApi.SocialNetWorkAdministration/Infrastructure/Mapping/UserProfile.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
-using Model.Models;
-using Model.Models.Dto;
+using DAL.Domain;
+using DAL.Domain.Dto;
 using System;
 using System.Collections.Generic;
 using System.Linq;
