+++ WebApi.SocialNetWorkAdministration/SeedDb.cs
@@ -1,9 +1,9 @@
-﻿using Model.Models;
+﻿using DAL.Domain;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using static Model.Models.User;
+using static DAL.Domain.User;
 
 namespace WebApi.SocialNetWorkAdministration
 {
